Refactor of fieldComponents and groupComponents
