+++ src/components/group/NormalForm.js
