+++ src/factory/BootstrapFactory.js
@@ -1,15 +1,15 @@
 import ComponentFactory from 'redux-autoform-utils/lib/factory/ComponentFactory';
-import TextBox from '../components/fieldComponents/TextBox';
-import Select from '../components/fieldComponents/Select';
-import TextArea from '../components/fieldComponents/TextArea';
-import Group from '../components/groupComponents/Group';
-import ArrayContainer from '../components/fieldComponents/ArrayContainer';
-import DateTimePicker from '../components/fieldComponents/DateTimePicker';
-import Lookup from '../components/fieldComponents/Lookup';
-import Static from '../components/fieldComponents/Static';
-import FieldGroup from '../components/fieldComponents/FieldGroup';
-import Checkbox from '../components/fieldComponents/CheckBox';
-import Radio from '../components/fieldComponents/Radio';
+import TextBox from '../components/field/TextBox';
+import Select from '../components/field/Select';
+import TextArea from '../components/field/TextArea';
+import Group from '../components/group/Group';
+import ArrayContainer from '../components/field/ArrayContainer';
+import DateTimePicker from '../components/field/DateTimePicker';
+import Lookup from '../components/field/Lookup';
+import Static from '../components/field/Static';
+import FieldGroup from '../components/field/FieldGroup';
+import Checkbox from '../components/field/CheckBox';
+import Radio from '../components/field/Radio';
 
 class BootstrapFactory extends ComponentFactory {
     constructor(config) {
